requirements.txt
pymysql>=1.0.2
pymongo>=4.0.0
python-dotenv>=0.21.0